requirements-dev.txt
# Dépendances de développement pour WaterBill API
# Installation: pip install -r requirements-dev.txt

# Tests
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0
factory-boy==3.3.3
coverage==7.10.7
Faker==37.8.0

# Formatage et linting
black==24.8.0
ruff==0.6.0

# Sécurité
bandit==1.7.9
safety==3.2.4
pip-audit==2.9.0

# Type checking
django-stubs==5.1.1
django-stubs-ext==5.2.5
djangorestframework-stubs==3.15.2
mypy-extensions==1.1.0

# Pre-commit
pre-commit==3.7.1

# Development utilities
identify==2.6.14
cfgv==3.4.0
nodeenv==1.9.1
virtualenv==20.34.0